+++ plugins/one_chat/lib/one_chat_web/views/shared_view.ex
@@ -172,7 +172,7 @@ defmodule OneChatWeb.SharedView do
   def has_role?(user, role, scope), do: Accounts.has_role?(user, role, scope)
   def has_role?(user, role), do: Accounts.has_role?(user, role)
 
-  def user_muted?(%{} = user, channel_id), do: OneChat.Channel.user_muted?(user.id, channel_id)
+  def user_muted?(%{} = user, channel_id), do: OneChat.Channel.user_muted?(channel_id, user.id)
 
   def content_home_title do
     "test"