+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * [UCX-3896] Message replacement patterns with OneChat.refresh_users_status exception
 * [UCX-3897] Fix Error when deleting user on system with only one admin.
+* [UCX-3900] Should not be able to mute a user in DM channel
 
 ## 1.0.0.beta11 (2018-04-05)